Used estimated partition count from RasterSummary object
S3GeoTiffRDD does not seem to propagate the estimated partition count after CutTiles stage
