+++ version.sbt
@@ -1,1 +1,1 @@
-version in ThisBuild := "0.7.0"
+version in ThisBuild := "0.8.0"
